+++ src/extension.ts
@@ -17,6 +17,7 @@ import { DemoPanel } from "./panels/DemoPanel";
 import { Preview } from "./preview/Preview";
 import { PresenterView } from "./presenterView/PresenterView";
 import { Config } from "./constants";
+import { PdfExportService } from "./services/PdfExportService";
 
 export async function activate(context: vscode.ExtensionContext) {
   Extension.getInstance(context);
@@ -34,10 +35,11 @@ export async function activate(context: vscode.ExtensionContext) {
   Preview.register();
   PresenterView.register();
   FileProvider.register();
-  Slides.registerCommands();
+  Slides.register();
   NotesService.registerCommands();
   DemoApi.register();
   UriHandler.register();
+  PdfExportService.register();
 
   console.log(`${Config.title} is active!`);
 }