+++ src/preview/components/MarkdownPreview.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { Messenger } from '@estruyf/vscode/dist/client/webview';
+import { messageHandler, Messenger } from '@estruyf/vscode/dist/client/webview';
 import { SlideLayout, WebViewMessages } from '../../constants';
 import { Markdown } from './Markdown';
 import { EventData } from '@estruyf/vscode';
@@ -40,16 +40,6 @@ export const MarkdownPreview: React.FunctionComponent<IMarkdownPreviewProps> = (
     }
   };
 
-  const slideClasses = React.useMemo(() => {
-    if (!layout) {
-      return '';
-    }
-
-    if (layout === SlideLayout.ImageLeft || layout === SlideLayout.ImageRight) {
-      return 'grid grid-cols-2 w-full h-full auto-rows-fr';
-    }
-  }, [layout]);
-
   const getBgStyles = React.useCallback(() => {
     if (!layout || layout === SlideLayout.ImageLeft || layout === SlideLayout.ImageRight) {
       return undefined;
@@ -75,54 +65,57 @@ export const MarkdownPreview: React.FunctionComponent<IMarkdownPreviewProps> = (
   }, []);
 
   return (
-    <div
-      key={crntFilePath}
-      ref={ref}
-      className={`slide fade-in ${theme || "default"} relative w-full h-full overflow-hidden`}
-      onMouseEnter={() => setShowControls(true)}
-      onMouseLeave={() => setShowControls(false)}
-      onMouseMove={handleMouseMove}
-      style={{ cursor: cursorVisible ? 'default' : 'none' }}
-    >
+    <>
+
       <div
-        className='slide__container absolute top-[50%] left-[50%] w-[960px] h-[540px]'
-        style={{ transform: 'translate(-50%, -50%) scale(var(--demotime-scale, 1))' }}>
+        key={crntFilePath}
+        ref={ref}
+        className={`slide fade-in ${theme || "default"} relative w-full h-full overflow-hidden`}
+        onMouseEnter={() => setShowControls(true)}
+        onMouseLeave={() => setShowControls(false)}
+        onMouseMove={handleMouseMove}
+        style={{ cursor: cursorVisible ? 'default' : 'none' }}
+      >
         <div
-          ref={slideRef}
-          className={`slide__layout ${slideClasses || ""} ${layout || "default"}`}
-          style={getBgStyles()}>
-          {
-            layout === SlideLayout.ImageLeft && (
-              <div className={`slide__image_left w-full h-full`} style={bgStyles}></div>
-            )
-          }
+          className='slide__container absolute top-[50%] left-[50%] w-[960px] h-[540px]'
+          style={{ transform: 'translate(-50%, -50%) scale(var(--demotime-scale, 1))' }}>
+          <div
+            ref={slideRef}
+            className={`slide__layout ${layout || "default"}`}
+            style={getBgStyles()}>
+            {
+              layout === SlideLayout.ImageLeft && (
+                <div className={`slide__image_left w-full h-full`} style={bgStyles}></div>
+              )
+            }
 
-          {
-            content && vsCodeTheme ? (
-              <div className='slide__content'>
-                <div className='slide__content__inner'>
-                  <Markdown
-                    content={content}
-                    vsCodeTheme={vsCodeTheme}
-                    webviewUrl={webviewUrl}
-                    updateTheme={setTheme}
-                    updateLayout={setLayout}
-                    updateBgStyles={setBgStyles}
-                  />
+            {
+              content && vsCodeTheme ? (
+                <div className='slide__content'>
+                  <div className='slide__content__inner'>
+                    <Markdown
+                      content={content}
+                      vsCodeTheme={vsCodeTheme}
+                      webviewUrl={webviewUrl}
+                      updateTheme={setTheme}
+                      updateLayout={setLayout}
+                      updateBgStyles={setBgStyles}
+                    />
+                  </div>
                 </div>
-              </div>
-            ) : null
-          }
+              ) : null
+            }
 
-          {
-            layout === SlideLayout.ImageRight && (
-              <div className={`slide__image_right w-full h-full`} style={bgStyles}></div>
-            )
-          }
+            {
+              layout === SlideLayout.ImageRight && (
+                <div className={`slide__image_right w-full h-full`} style={bgStyles}></div>
+              )
+            }
+          </div>
         </div>
-      </div>
 
-      <SlideControls show={showControls && cursorVisible} path={crntFilePath} />
-    </div>
+        <SlideControls show={showControls && cursorVisible} path={crntFilePath} />
+      </div>
+    </>
   );
 };